made separate file for component and cleaned up code to include math instead of if statments
